Renaming the repository
