+++ include/ChessLib/Defines.hpp
