+++ include/ChessLib/Figure.hpp
