+++ include/ChessLib/Figures.hpp
