+++ src/Move.cpp
@@ -1,4 +1,4 @@
-#include "Chess/Move.hpp"
+#include "ChessLib/Move.hpp"
 
 namespace Chess
 {
